Ensures a meaningful StatsD bucket is created
Since .by_page_and_metric already sets a base StatsD key, we cannot delegate the checking for the presence of metric to StatsD.

To ensure that we have a properly namespaced StatsD key, we need to check for metric being present in this method.

We can delegate that checking to StatsD for duration, though.